backend/requirements.txt
# Backend
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database
pymongo==4.6.0

# Deep Learning
tensorflow==2.17.1
keras==3.3.3
torch==2.2.2
torchvision==0.17.2
ultralytics>=8.2.0

# Image Processing
opencv-python-headless==4.8.1.78
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.4